requirements.txt
# local package
-e .

# external requirements
flake8==3.6.0
jupyter==1.0.0
matplotlib==2.1.2
numpy==1.15.4
pandas==0.22.0
pydicom==1.2.0
pynrrd==0.3.5
scikit-image==0.14.1
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
tensorboard==1.12.0
tensorflow-gpu==1.12.0
